feat(db): Phase C — enable PostgreSQL backend at runtime
Complete the multi-database effort: PostgreSQL is now a runtime-selectable
backend alongside SQLite (chosen once at startup from DATABASE_URL). SQLite
stays the zero-config default; every SQLite path is unchanged.

All dialect differences are isolated behind `Dialect` / a `pg_rewrite` shim so
model SQL is still written once:

- Pin every PG connection to `TimeZone=UTC` so timestamp-string cursors stay
  byte-identical to SQLite's `datetime('now')` TEXT.
- Central `pg_rewrite` shim rewrites `datetime('now')`→`now()` in the dispatch
  macros and dynamic-exec helpers.
- Composite cursor, snapshot `read_after`, and neighbor comparisons wrap the
  column in `to_char(...,'YYYY-MM-DD HH24:MI:SS')` on PG (`Dialect::cursor_ts`)
  so the string-ordered cursor is preserved; `fetch_sort_ts` / neighbor sort_ts
  reads use the same.
- Interval arithmetic forks to `make_interval` (`Dialect::days_ago`); date-range
  bounds bind as `NaiveDate` (PG implicitly casts date→timestamptz).
- Timestamp columns read into `String` wrap in `to_char`; `DATE()`→`to_char`,
  `PRAGMA`→`pg_database_size()`, epoch via `EXTRACT(EPOCH …)`.
- Quote the reserved `"user"` table uniformly (valid on both backends).
- PG migration: non-id `INTEGER`→`BIGINT` and `has_icon` `CAST(... AS BIGINT)`
  to match the `i64` reads under sqlx-PG's strict int widths.
- Drop the config `validate()` PostgreSQL rejection.

Validated against live PostgreSQL 17 by the env-gated `tests/postgres_test.rs`
(skips without TEST_DATABASE_URL) plus a CI Postgres service lane. Full SQLite
suite (975) and the PG integration test both pass.

Deferred (TODO(phase-d)): the PG cursor `to_char` predicate is not sargable;
bind the cursor as timestamptz to restore index range scans.

+++ migrations/postgres/0001_initial.sql
@@ -50,7 +50,7 @@ CREATE TABLE IF NOT EXISTS feed (
     custom_user_agent TEXT,
     http2_disabled BOOLEAN NOT NULL DEFAULT FALSE,
     custom_referrer TEXT,
-    bucket INTEGER,
+    bucket BIGINT,
     created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
     updated_at TIMESTAMPTZ NOT NULL DEFAULT now(),
     UNIQUE(category_id, url)
@@ -130,10 +130,10 @@ CREATE INDEX IF NOT EXISTS idx_image_entity ON image(entity_type, entity_id);
 CREATE TABLE IF NOT EXISTS user_settings (
     id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     user_id BIGINT NOT NULL UNIQUE REFERENCES "user"(id) ON DELETE CASCADE,
-    entries_per_page INTEGER NOT NULL DEFAULT 30,
+    entries_per_page BIGINT NOT NULL DEFAULT 30,
     save_services TEXT,
     theme TEXT,
-    retention_read_days INTEGER NOT NULL DEFAULT 0,
+    retention_read_days BIGINT NOT NULL DEFAULT 0,
     created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
     updated_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
